requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.5
pinecone>=3.0.0
google-api-python-client>=2.110.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
python-docx>=1.1.0
PyPDF2>=3.0.1
openai>=1.10.0
python-dotenv>=1.0.0
fastapi>=0.108.0
uvicorn>=0.25.0
pydantic>=2.5.0
requests>=2.31.0
